Copy relevancy from kw18 column 19 up to track in detection mode
The track_oracle data model of a track is

  [track-data]
      +------- [ frame 000 data ]
      +------- [ frame 001 data ]
      ...
      +------- [ frame NNN data ]

When the relevancy is read out of the 19th column in a kw18, that
value lives at the frame level. But the scoring code is looking for
it at the track level. Enabling detection mode transforms a track
such as above into

  [track-data]
      +------- [ frame 000 data ]
  [track-data]
      +------- [ frame 001 data ]
      ...
  [track-data]
      +------- [ frame NNN data ]

...but the relevancy wasn't copied up to the track. This patch fixes
that.